Refactor pokemon to share same base interface Used to funnel pokemon randomization through the same set of methods Add "NextElement" to RNG methods

+++ Common/PokemonDefinitions/Colosseum/ColGiftPokemon.cs
@@ -52,7 +52,7 @@ namespace XDCommon.PokemonDefinitions
 			_ => PikachuOffset
 		};
 
-        public byte Index { get; }
+        public int Index { get; }
         public ushort Exp => 0;
 
         public ushort[] Moves => new ushort[Constants.NumberOfPokemonMoves];
